renamed src to lib because not all source files (circuits and contracts) are in there
