+++ lib/merkleTree.ts
